Investigations created

+++ app/src/main/java/db/entity/AnswerData.kt
@@ -19,10 +19,10 @@ import androidx.room.*
 data class AnswerData(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = "id")
-    val id: Int,
+    val id: Long,
     @ColumnInfo(name = "answerName") val answerName: String,
     @ColumnInfo(name = "answer") val answer: String,
-    @ColumnInfo(name = "questionDataId") val questionDataId: Int
+    @ColumnInfo(name = "questionDataId") val questionDataId: Long
 ) {
 
 }
